reorganize project as a python package
