+++ README.md
@@ -1,3 +1,13 @@
 # mlb-betting-model
 
 ![2023-03-29 11_02_36-Untitled _ Visual Paradigm Online - Work - Microsoft​ Edge](https://user-images.githubusercontent.com/29851231/228599081-278a6b6e-898f-4a5c-b8a5-7887da673ac8.png)
+
+
+## Running python files
+
+Since I made this project into a proper python package, you'll have to run the files a different way than normal. To execute `batter_preprocessing.py`:
+
+```sh
+cd src
+python -m mlbbetting.preprocessing.batter_preprocessing
+```
