+++ src/mlbbetting/weather/main.py
@@ -1,10 +1,11 @@
-import routines as weather_routines
+from . import routines as weather_routines
 import json
-
 
 def test():
     current_weather = weather_routines.get_weather_by_zip('60614')
     print(json.dumps(current_weather.to_dict(), indent=4))
 
-test()
 
+# Main logic
+if __name__ == "__main__":
+    test()
